Use a custom SyncWriterWrapper
